requirements.txt
selenium==4.33.0
pytest==8.3.5
python-dotenv==1.1.0
